Publish translation files to GitHub Pages
The Drupal translation system will only save the status of a translation
if the translation is provided with a timestamp as seen in 
locale_translation_status_save().

When using raw files from GitHub through raw.githubusercontent.com files
are returned without a Last-Modified header. Consequently the timestamp
for these files is set to 0.

Try to avoid this by publishing updated translations to GitHub Pages.
Hopefully this will provide a better basis for returning appropriate
HTTP headers.

Currently the project does not use GitHub Pages for other purposes but
it might in the future. Publish to a subdirectory of the repo to make 
room for other content.